[sim][world_generation] Separated test.py into 3 logical files
